Create code.py
We have 1st cleaned the data by splitting it into train and test . we have removed outliers by plotting the scatter plot for each and every attribute. For numerical values we have replaced with the median values and for categorical values have have replaces null places with mode. Then we used label encoder for categorical values. At last we used pca to determine the error and visualized it with graph